Implement 2 of 4 endings
