--- Assets/MenuUtils.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1c932c3f551732e43b012fe8291cc3b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
