+++ Assets/Scripts/ImageFader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04ce40f6f2b6a6341ab5e278af99b571
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
